python-projects-final/requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
sentence-transformers
python-dotenv
beautifulsoup4
